pyproject.toml
[project]
name = "ownaichatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.11.0",
    "langchain>=0.3.26",
    "langchain-community>=0.3.26",
    "langchain-openai>=0.3.24",
    "pypdf>=5.6.1",
    "python-dotenv>=1.1.0",
    "streamlit>=1.46.0",
]
